Nest obligatory's Grpc, JsonRpc and Sse satellites
`GrpcChannel`, `GrpcError` and `GrpcFraming` become `Grpc.Channel`, `Grpc.Error`
and `Grpc.Framing`; `JsonRpcError` becomes `JsonRpc.Error`; `SseError` and
`SseSource` become `Sse.Error` and `Sse.Source`. Six names leave the umbrella's
exported surface.

Two hazards the proposal predicted both fired here. `grpcInternal`'s quoted
expansions carry anonymous `Tactic[GrpcError]` and `Tactic[Http2.Error]` givens,
which synthesised the same name once the first became `Tactic[Grpc.Error]`; they
are now named `grpcTactic` and `http2Tactic`. And `obligatory.JsonRpc.scala` ends
its imports with `httpBackends.virtualMachine`, a choice import that must follow
`telekinesis.*` — new imports go into the alphabetical run above it, never
re-sorted over it.
