+++ lib/obligatory/src/grpc/soundness_obligatory_grpc.scala
@@ -32,4 +32,4 @@
                                                                                                   */
 package soundness
 
-export obligatory.{Grpc, GrpcChannel, GrpcError, GrpcFraming, GrpcSessional}
+export obligatory.{Grpc, GrpcSessional}
