--- lib/obligatory/src/json/obligatory.JsonRpcError.scala
@@ -1,58 +1,0 @@
-                                                                                                  /*
-┏━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━┓
-┃                                                                                                  ┃
-┃                                                   ╭───╮                                          ┃
-┃                                                   │   │                                          ┃
-┃                                                   │   │                                          ┃
-┃   ╭───────╮╭─────────╮╭───╮ ╭───╮╭───╮╌────╮╭────╌┤   │╭───╮╌────╮╭────────╮╭───────╮╭───────╮   ┃
-┃   │   ╭───╯│   ╭─╮   ││   │ │   ││   ╭─╮   ││   ╭─╮   ││   ╭─╮   ││   ╭─╮  ││   ╭───╯│   ╭───╯   ┃
-┃   │   ╰───╮│   │ │   ││   │ │   ││   │ │   ││   │ │   ││   │ │   ││   ╰─╯  ││   ╰───╮│   ╰───╮   ┃
-┃   ╰───╮   ││   │ │   ││   │ │   ││   │ │   ││   │ │   ││   │ │   ││   ╭────╯╰───╮   │╰───╮   │   ┃
-┃   ╭───╯   ││   ╰─╯   ││   ╰─╯   ││   │ │   ││   ╰─╯   ││   │ │   ││   ╰────╮╭───╯   │╭───╯   │   ┃
-┃   ╰───────╯╰─────────╯╰────╌╰───╯╰───╯ ╰───╯╰────╌╰───╯╰───╯ ╰───╯╰────────╯╰───────╯╰───────╯   ┃
-┃                                                                                                  ┃
-┃    Soundness, version 0.64.0.                                                                    ┃
-┃    © Copyright 2021-25 Jon Pretty, Propensive OÜ.                                                ┃
-┃                                                                                                  ┃
-┃    The primary distribution site is:                                                             ┃
-┃                                                                                                  ┃
-┃        https://soundness.dev/                                                                    ┃
-┃                                                                                                  ┃
-┃    Licensed under the Apache License, Version 2.0 (the "License"); you may not use this file     ┃
-┃    except in compliance with the License. You may obtain a copy of the License at                ┃
-┃                                                                                                  ┃
-┃        https://www.apache.org/licenses/LICENSE-2.0                                               ┃
-┃                                                                                                  ┃
-┃    Unless required by applicable law or agreed to in writing,  software distributed under the    ┃
-┃    License is distributed on an "AS IS" BASIS,  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND,    ┃
-┃    either express or implied. See the License for the specific language governing permissions    ┃
-┃    and limitations under the License.                                                            ┃
-┃                                                                                                  ┃
-┗━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━┛
-                                                                                                  */
-package obligatory
-
-import anticipation.*
-import fulminate.*
-import vacuous.*
-
-object JsonRpcError:
-  enum Reason(val number: Int) extends Clarification:
-    case UnknownMethod extends Reason(1)
-    case Failed        extends Reason(2)
-    case Abandoned     extends Reason(3)
-
-  given communicable: Reason is Communicable =
-    case Reason.UnknownMethod => m"the method name was not recognised by the dispatcher"
-    case Reason.Failed        => m"the peer answered the request with an error"
-    case Reason.Abandoned     => m"the request was never answered"
-
-// `code` and `detail` carry the peer's `error.code` and `error.message` when the fault came back
-// over the wire, so a caller can map the failure onto its own protocol's vocabulary — an LSP
-// client, for instance, recovers an `Lsp.Error.Reason` from the code.
-case class JsonRpcError
-   ( reason: JsonRpcError.Reason,
-     code:   Optional[Int]  = Unset,
-     detail: Optional[Text] = Unset )
-   ( using Diagnostics )
-extends Error(721, reason.number)(m"the JSON-RPC operation failed because $reason")
