--- lib/obligatory/src/json/obligatory.SseSource.scala
@@ -1,68 +1,0 @@
-                                                                                                  /*
-┏━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━┓
-┃                                                                                                  ┃
-┃                                                   ╭───╮                                          ┃
-┃                                                   │   │                                          ┃
-┃                                                   │   │                                          ┃
-┃   ╭───────╮╭─────────╮╭───╮ ╭───╮╭───╮╌────╮╭────╌┤   │╭───╮╌────╮╭────────╮╭───────╮╭───────╮   ┃
-┃   │   ╭───╯│   ╭─╮   ││   │ │   ││   ╭─╮   ││   ╭─╮   ││   ╭─╮   ││   ╭─╮  ││   ╭───╯│   ╭───╯   ┃
-┃   │   ╰───╮│   │ │   ││   │ │   ││   │ │   ││   │ │   ││   │ │   ││   ╰─╯  ││   ╰───╮│   ╰───╮   ┃
-┃   ╰───╮   ││   │ │   ││   │ │   ││   │ │   ││   │ │   ││   │ │   ││   ╭────╯╰───╮   │╰───╮   │   ┃
-┃   ╭───╯   ││   ╰─╯   ││   ╰─╯   ││   │ │   ││   ╰─╯   ││   │ │   ││   ╰────╮╭───╯   │╭───╯   │   ┃
-┃   ╰───────╯╰─────────╯╰────╌╰───╯╰───╯ ╰───╯╰────╌╰───╯╰───╯ ╰───╯╰────────╯╰───────╯╰───────╯   ┃
-┃                                                                                                  ┃
-┃    Soundness, version 0.64.0.                                                                    ┃
-┃    © Copyright 2021-25 Jon Pretty, Propensive OÜ.                                                ┃
-┃                                                                                                  ┃
-┃    The primary distribution site is:                                                             ┃
-┃                                                                                                  ┃
-┃        https://soundness.dev/                                                                    ┃
-┃                                                                                                  ┃
-┃    Licensed under the Apache License, Version 2.0 (the "License"); you may not use this file     ┃
-┃    except in compliance with the License. You may obtain a copy of the License at                ┃
-┃                                                                                                  ┃
-┃        https://www.apache.org/licenses/LICENSE-2.0                                               ┃
-┃                                                                                                  ┃
-┃    Unless required by applicable law or agreed to in writing,  software distributed under the    ┃
-┃    License is distributed on an "AS IS" BASIS,  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND,    ┃
-┃    either express or implied. See the License for the specific language governing permissions    ┃
-┃    and limitations under the License.                                                            ┃
-┃                                                                                                  ┃
-┗━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━┛
-                                                                                                  */
-package obligatory
-
-import anticipation.*
-import contingency.*
-import prepositional.*
-import rudiments.*
-import spectacular.*
-import turbulence.*
-import zephyrine.*
-import vacuous.*
-
-class SseSource(capacity: Int):
-  private val mutex: Mutex = Mutex()
-  @scala.caps.unsafe.untrackedCaptures
-  private val buffer: scala.Array[Sse] = new scala.Array(capacity)
-
-  @scala.caps.unsafe.untrackedCaptures
-  private var current: Int = 0
-  @scala.caps.unsafe.untrackedCaptures
-  private var spool: Relay[Sse] = Relay()
-
-  def put[entity: Encodable in Sse](value: entity): Unit = mutex:
-    val sse = value.encode.copy(id = current.show)
-    buffer.asInstanceOf[scala.Array[Sse]^](current%capacity) = sse
-    spool.put(sse)
-    current += 1
-
-  def stream(start: Optional[Int] = Unset): Chain[Sse] raises SseError = mutex:
-    start.let: start =>
-      spool.stop()
-      spool = Relay()
-
-      if current - start - 1 > capacity then abort(SseError(SseError.Reason.CapacityExceeded)) else
-        ((start + 1) until current).map: index => spool.put(buffer(index%capacity))
-
-    Chain.from(spool.stream.records)
