+++ lib/obligatory/src/json/soundness_obligatory_json.scala
@@ -32,4 +32,4 @@
                                                                                                   */
 package soundness
 
-export obligatory.{JsonRpc, JsonRpcError, remote, Sse, SseError, SseSource}
+export obligatory.{JsonRpc, remote, Sse}
